Move shell quoting out of pkg/client/logging in to their own pkg/shellquote
