+++ .golangci.yml
@@ -33,7 +33,7 @@ linters-settings:
       - github.com/golang/protobuf: "Use `google.golang.org/protobuf` instead of `github.com/golang/protobuf`"
       - github/datawire/dlib/dutil: "Use either `github.com/datawire/dlib/derror` or `github.com/datawire/dlib/dhttp` instead of `github.com/datawire/dlib/dutil`"
       - golang.org/x/net/http2/h2c: "Use `github.com/datawire/dlib/dhttp` instead of `golang.org/x/net/http2/h2c`"
-      - github.com/kballard/go-shellquote: "Use `github.com/telepresenceio/telepresence/pkg/client/logging.ShellString` instead of `github.com/kballard/go-shellquote.Join`"
+      - github.com/kballard/go-shellquote: "Use `github.com/telepresenceio/telepresence/pkg/shellquote.ShellString` instead of `github.com/kballard/go-shellquote.Join`"
   gomodguard:
     blocked:
       modules:
